Module Tezos_benchmarks_proto_020_PsParisC.Cache_benchmarks

val ns : Tezos_benchmark.Namespace.cons

val fv : string -> Tezos_benchmark.Free_variable.t

Alpha_context.Cache-related benchmarks

val assert_ok_lwt : ('a, 'b) Stdlib.result Lwt.t -> 'a

val assert_ok : ('a, 'b) Stdlib.result -> 'a

module Admin = Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Cache.Admin

module Cache = Tezos_protocol_020_PsParisC.Protocol.Script_cache

We can't construct a dummy cache client from outside the protocol. We'll have to benchmark the Environment_cache through the interface exposed by Script_cache.

Constructing a dummy cached value

val make_context : rng_state:Stdlib.Random.State.t -> Tezos_raw_protocol_020_PsParisC__Alpha_context.t

val throwaway_context : Tezos_raw_protocol_020_PsParisC__Alpha_context.t

val dummy_script : Cache.cached_contract

Creating dummy cache value identifiers.

module Cache_shared_config : sig ... end

Configuration shared among all cache benchmarks.

val identifier_of_contract : Tezos_protocol_020_PsParisC.Protocol.Contract_hash.t -> Cache.identifier

val contract_of_int : int -> Tezos_protocol_020_PsParisC.Protocol.Contract_hash.t

val identifier_of_int : int -> Cache.identifier

val prepare_context : Stdlib.Random.State.t -> int -> Tezos_raw_protocol_020_PsParisC.Alpha_context.context * Cache.identifier

Prepare a context with a cache of the prescribed cardinality. A key in the domain of the cache is returned along the context: this key is used to benchmark (successful) cache accesses.

module Cache_update_benchmark : Benchmarks_proto.Benchmark.S

Benchmark Script_cache.update. This almost directly calls Environment_cache.update. We also use the result of this benchmark to assign a cost to Environment_cache.find, which alas can't be directly benchmarked from the interface provided by Script_cache.