Module Global_constants_storage_expand_models.Global_constants_storage_expand_no_constant_branch

val name : Tezos_benchmark.Namespace.t

Name of the benchmark

val info : string

Description of the benchmark

val module_filename : string

Filename of the benchmark module

val group : Benchmarks_proto.Benchmark.group

Inference group of the benchmark

val tags : string list

Tags of the benchmark

type config

Configuration of the benchmark (eg sampling parameters, paths, etc)

val default_config : config

Default configuration of the benchmark

val config_encoding : config Tezos_base.TzPervasives.Data_encoding.t

Configuration encoding

type workload

Benchmark workload

val workload_encoding : workload Tezos_base.TzPervasives.Data_encoding.t

Workload encoding

val workload_to_vector : workload -> Tezos_benchmark.Sparse_vec.String.t

Optional conversion to vector, for report generation purposes

val model : name:Tezos_benchmark.Namespace.t -> workload Tezos_benchmark.Model.t

Cost model

val purpose : Tezos_benchmarks_proto_020_PsParisC.Benchmarks_proto.Benchmark_base.purpose

Benchmark purpose. The developer can submit an explanation as to what the benchmark is for when it's not generating code (i.e. in the case of Other_purpose).

val create_benchmark : rng_state:Stdlib.Random.State.t -> config -> workload Tezos_benchmark.Generator.benchmark

Creates a benchmark, ready to be run. The benchmarks are thunked to prevent evaluating the workload until needed.