Module Tezos_benchmarks_proto_020_PsParisC.Sc_rollup_benchmarks

val ns : Tezos_benchmark.Namespace.cons

val fv : string -> Tezos_benchmark.Free_variable.t

val (--) : int -> int -> Tezos_benchmark.Base_samplers.range

module Pvm_state_generator : sig ... end

This section contains preliminary definitions for building a pvm state from scratch.

module Sc_rollup_verify_output_proof_benchmark : sig ... end

This benchmark estimates the cost of verifying an output proof for the Wasm PVM. The inferred cost model is c1 + c2 * proof_length.

module Sc_rollup_deserialize_output_proof_benchmark : sig ... end

This benchmark estimates the cost of verifying an output proof for the Wasm PVM. The inferred cost model is c1 + c2 * proof_length.

module Sc_rollup_install_boot_sector_benchmark : sig ... end

This benchmark estimates the cost of installing a boot sector.