Module Tezos_benchmarks_proto_020_PsParisC.Skip_list_benchmarks

This module provides benchmarks for skip list operations for basis = 4.

module Skip_list : sig ... end

val ns : Tezos_benchmark.Namespace.cons

val fv : string -> Tezos_benchmark.Free_variable.t

module Next : Benchmarks_proto.Benchmark.S

Benchmark for the Skip_list_repr.next function. It is used for estimating the parameters for Skip_list_cost_model.model_next.

module Hash_cell : Benchmarks_proto.Benchmark.S

Benchmark for the Sc_rollup_inbox_repr.hash_skip_list_cell function. It is used for estimating the parameters for Skip_list_cost_model.model_hash_cell. The model estimates hashing a skip_list cell content and all its back pointers.